Code for Input validation for Insert page #12
Added validation using jquery validation plug in.
It works when updating and inserting.
Decided to keep the code validation as is working right now, and not
include it in a "remote" validation, because that implied that we had to
change the output of the api and I'm not sure it's worthwhile.